requirements.txt
notepad requirements.txt
```
Paste this:
```
flask==3.1.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.48
pyodbc==5.3.0
python-dotenv==1.2.2
reportlab==4.4.10
requests==2.32.5
groq==0.23.1
gunicorn==21.2.0